requirements.txt
Package          Version
---------------- -------
beautifulsoup4   4.7.1  
EasyProcess      0.2.7  
et-xmlfile       1.0.1  
jdcal            1.4.1  
openpyxl         2.6.2  
Pillow           6.0.0  
pip              10.0.1 
pytesseract      0.2.6  
PyVirtualDisplay 0.2.3  
selenium         3.141.0
setuptools       39.0.1 
six              1.12.0 
soupsieve        1.9.1  
tqdm             4.32.1 
urllib3          1.25.3 
